requirements.txt
tweepy==3.10.0
# psycopg2-binary==2.8.6
pika==1.2.0
PyYAML==5.4.1
feedparser==6.0.10
openai==0.27.8
beautifulsoup4==4.12.2
requests~=2.31.0
pydantic~=2.0.3
schedule~=1.2.0